+++ src/api/admin/types.ts
@@ -792,6 +792,16 @@ export type StorageInfoResponse = StorageInfo;
 /**
  * @public
  */
+export type DumpNetworkStatsRequest = void;
+
+/**
+ * @public
+ */
+export type DumpNetworkStatsResponse = string;
+
+/**
+ * @public
+ */
 export interface AdminApi {
   attachAppInterface: Requester<
     AttachAppInterfaceRequest,
@@ -827,4 +837,8 @@ export interface AdminApi {
     GrantZomeCallCapabilityResponse
   >;
   storageInfo: Requester<StorageInfoRequest, StorageInfoResponse>;
+  dumpNetworkStats: Requester<
+    DumpNetworkStatsRequest,
+    DumpNetworkStatsResponse
+  >;
 }