+++ CHANGELOG.md
@@ -4,9 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## \[Unreleased\]
 
-- Add new conductor API endpoint `storageInfo` 
-
 ### Added
+- New Admin API endpoint `storageInfo`.
 ### Changed
 ### Fixed
 ### Removed